requirements.txt
# ============================================
# Power Interpreter MCP - Requirements
# ============================================

# --- Core Framework ---
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
httpx==0.28.1

# --- Database ---
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10

# --- MCP Protocol ---
mcp==1.2.0

# --- Data Analysis (Pre-installed for sandbox) ---
pandas==2.2.3
numpy==2.2.1
openpyxl==3.1.5
xlsxwriter==3.2.0
pdfplumber==0.11.4
tabulate==0.9.0

# --- Visualization ---
matplotlib==3.10.0
plotly==5.24.1
seaborn==0.13.2

# --- Statistics & ML ---
scipy==1.15.1
scikit-learn==1.6.1
statsmodels==0.14.4

# --- Utilities ---
python-dotenv==1.0.1
python-dateutil==2.9.0
aiofiles==24.1.0
orjson==3.10.14
rich==13.9.4
